langgraph/requirements.txt
openai
python-dotenv
duckduckgo-search
pypdf
lancedb 
tantivy
streamlit 
pandas
yfinance
openai 
duckduckgo-search 
yfinance 
sqlalchemy 
fastapi[standard]
langgraph
langchain
langchain_groq
langchain_openai
langgraph-cli[inmem]
langchain_core
langchain-community
faiss-cpu
langdetect